Add AnnotateRoutes::Helpers.extract_magic_comments_from_array
